backend/requirements.txt
torch
torchvision
flask
flask-cors
Pillow
ultralytics
